Browser support for sqrl-redis-functions
