+++ packages/sqrl-common/src/index.ts
@@ -7,6 +7,7 @@ export * from "./RenderedSpan";
 export * from "./SqrlBoxed";
 export * from "./SqrlObject";
 
+export * from "./arrayToHex";
 export * from "./bufferToHexEncodedAscii";
 export * from "./compare";
 export * from "./emptyFunction";
