+++ packages/sqrl-redis-functions/package.json
@@ -24,6 +24,9 @@
   ],
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
+  "browser": {
+    "./lib/services/RedisService.js": "./lib/browser/RedisService.js"
+  },
   "scripts": {
     "clean": "rimraf lib && rimraf coverage",
     "format": "prettier --write \"{src,__tests__}/**/*.ts\"",
@@ -57,9 +60,9 @@
     "fast-stable-stringify": "^1.0.0",
     "ioredis": "^4.19.1",
     "moment": "^2.24.0",
-    "murmurhash-native": "^3.4.1",
-    "sqrl": "^0.6.4",
-    "sqrl-common": "^0.6.4"
+    "murmur-128": "^0.2.1",
+    "sqrl": "^0.6.6-beta.4",
+    "sqrl-common": "^0.6.6-beta.4"
   },
   "gitHead": "d6f4789442ac2e1dea7b1f39b87c6db3395fd815"
 }
