+++ packages/sqrl/src/jslib/murmurhashJson.ts
@@ -6,20 +6,7 @@
 import murmur128x64 = require("murmur-128");
 import stringify = require("fast-stable-stringify");
 import { TextEncoder } from "util";
-
-const hexDigits = [];
-for (let i = 0; i < 256; i++) {
-    const hexOctet = i.toString(16).padStart(2, "0");
-    hexDigits.push(hexOctet);
-}
-
-export function arrayToHex(data: Uint8Array): string {
-  const digits = new Array(data.byteLength);
-  for (let i = 0; i < data.length; i++) {
-      digits.push(hexDigits[data[i]]);
-  }
-  return digits.join("");
-}
+import {arrayToHex} from "sqrl-common";
 
 
 export function murmurhashJsonHexSync(input: any): string {
